+++ lib/rat_error/formatter.ex
@@ -1,0 +1,9 @@
+defmodule RatError.Formatter do
+  @moduledoc """
+  """
+
+  @doc """
+  """
+  def format(_a, _b, _c, _d) do
+  end
+end