+++ test/rat_error_test.exs
@@ -1,8 +1,5 @@
 defmodule RatErrorTest do
   use ExUnit.Case
+  use RatError
   doctest RatError
-
-  test "greets the world" do
-    assert RatError.hello() == :world
-  end
 end